libp2p_identify

Struct Behaviour

pub struct Behaviour { /* private fields */ }

Network behaviour that automatically identifies nodes periodically, returns information about them, and answers identify queries from other nodes.

All external addresses of the local node supposedly observed by remotes are reported via ToSwarm::NewExternalAddrCandidate.

Implementations

impl Behaviour

pub fn new(config: Config) -> Self

Creates a new identify Behaviour.

pub fn push<I>(&mut self, peers: I)
where I: IntoIterator<Item = PeerId>,

Initiates an active push of the local peer information to the given peers.