Type Alias libp2p_identify::IdentifyConfig

pub type IdentifyConfig = Config;

👎Deprecated since 0.40.0: Use re-exports that omit Identify prefix, i.e. libp2p_identify::Config

Aliased Type

struct IdentifyConfig {
    pub protocol_version: String,
    pub local_public_key: PublicKey,
    pub agent_version: String,
    pub initial_delay: Duration,
    pub interval: Duration,
    pub push_listen_addr_updates: bool,
    pub cache_size: usize,
}

Fields
protocol_version: String

Application-specific version of the protocol family used by the peer, e.g. ipfs/1.0.0 or polkadot/1.0.0.

local_public_key: PublicKey

The public key of the local node. To report on the wire.

agent_version: String

Name and version of the local peer implementation, similar to the User-Agent header in the HTTP protocol.

Defaults to rust-libp2p/<libp2p-identify-version>.

initial_delay: Duration

The initial delay before the first identification request is sent to a remote on a newly established connection.

Defaults to 500ms.

interval: Duration

The interval at which identification requests are sent to the remote on established connections after the first request, i.e. the delay between identification requests.

Defaults to 5 minutes.

push_listen_addr_updates: bool

Whether new or expired listen addresses of the local node should trigger an active push of an identify message to all connected peers.

Enabling this option can result in connected peers being informed earlier about new or expired listen addresses of the local node, i.e. before the next periodic identify request with each peer.

Disabled by default.

cache_size: usize

How many entries of discovered peers to keep before we discard the least-recently used one.

Disabled by default.