Type Alias libp2p_identify::IdentifyEvent

pub type IdentifyEvent = Event;

👎Deprecated since 0.40.0: Use re-exports that omit Identify prefix, i.e. libp2p_identify::Event

Aliased Type

enum IdentifyEvent {
    Received {
        peer_id: PeerId,
        info: Info,
    },
    Sent {
        peer_id: PeerId,
    },
    Pushed {
        peer_id: PeerId,
    },
    Error {
        peer_id: PeerId,
        error: ConnectionHandlerUpgrErr<UpgradeError>,
    },
}

Variants

Received

Fields

peer_id: PeerId

The peer that has been identified.

info: Info

The information provided by the peer.

Identification information has been received from a peer.

Sent

Fields

peer_id: PeerId

The peer that the information has been sent to.

Identification information of the local node has been sent to a peer in response to an identification request.

Pushed

Fields

peer_id: PeerId

The peer that the information has been sent to.

Identification information of the local node has been actively pushed to a peer.

Error

Fields

peer_id: PeerId

The peer with whom the error originated.

error: ConnectionHandlerUpgrErr<UpgradeError>

The error that occurred.

Error while attempting to identify the remote.